+++ opentelemetry-sdk/CHANGELOG.md
@@ -6,6 +6,8 @@
   ([#1425](https://github.com/open-telemetry/opentelemetry-python/pull/1425))
 - Add local/remote samplers to parent based sampler
   ([#1440](https://github.com/open-telemetry/opentelemetry-python/pull/1440))
+- Add `fields` to propagators
+  ([#1374](https://github.com/open-telemetry/opentelemetry-python/pull/1374))
 
 ## Version 0.16b0